Worker/requirements.txt
autopep8==2.3.2
esptool==5.1.0
mpremote==1.26.1
paho-mqtt==2.1.0
pycodestyle==2.14.0
pylint==4.0.3
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
